Quickly patching the code around chunks to work with 64 bit words instead of a byte

+++ include/lit/vm/lit_object.h
@@ -223,7 +223,7 @@ typedef struct {
 	LitFunction* function;
 	LitClosure* closure;
 
-	uint8_t* ip;
+	uint64_t* ip;
 	LitValue* slots;
 
 	bool result_ignored;
